Some fixes to controllers and migrations

+++ database/migrations/2018_04_26_090339_create_showings_table.php
@@ -19,7 +19,7 @@ class CreateShowingsTable extends Migration
             $table->dateTime('showingTime');
             $table->dateTime('showingEndTime');
             $table->boolean('active')->default(true);
-            $table->decimal('price', 8, 2);
+            $table->bigInteger('price');
             $table->timestamps();
             $table->softDeletes();
